实现接口原型 ipicserver_0.0.1
实现接口原型 ipicserver_0.0.1

+++ server/picStoryServer/src/main/java/com/towne/framework/hibernate/dao/IDao.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface IDao<T> {
 	
 	public void add(Object object);
-	public void delete(int id);
+	public void delete(long id);
 	public void update(Object object);
 	public List<T> query(String queryString);
-	public T findById(int id);
+	public T findById(long id);
 	
 }
